Added more user input
I created prompt so the player can chose their input instead of assigning it to playerSelection every round. The return string inside playRound() has been simplified and the selections are uppercase so the user input can be case insensitive. I also added a string what tells you what you and the computer have choosen. There is also a score displayed in the console.